chore: exported getUserRole
